+++ NEWS.md
@@ -18,6 +18,7 @@
 * Added `setup_logging` to enable users to specify the level and location of logging (wrapping functionality from `futile.logger`). 
 * Updated `discretised_gamma_pmf` (discretised truncated Gamma PMF) to constrain gamma shape and (inverse) scale parameters to be positive and finite (`alpha > 0` and `beta > 0`).
 * Fixed `readLines` incomplete final line warnings.
+* Implemented progress bar support using `progressr`.
 
 # EpiNow2 1.1.0